added new welcome screen and unlock password

+++ views/UserMembershipView.py
@@ -1,5 +1,3 @@
-import cv2
-import face_recognition
 from kivy.properties import ObjectProperty
 from kivy.uix.popup import Popup
 from kivy.uix.relativelayout import RelativeLayout
@@ -22,7 +20,6 @@ except ImportError:
 import string
 from manage.rpi_face_recon import *
 from kivy.clock import Clock
-from views.GlobalComponents import ConfirmationDialogContent
 
 class UserMembershipView(MDScreen):
     found_user = ObjectProperty()
@@ -34,20 +31,20 @@ class UserMembershipView(MDScreen):
         print(f"Welcome {self.found_user[0]}, {self.found_user[1]} !")
         self.ids.topbar.title = f"Welcome {self.found_user[0]}, {self.found_user[1]} !"
         self.ids.user_welcome_screen.found_user = self.found_user
-        self.ids.update_pin_screen.found_user = self.found_user
+        self.ids.update_password_screen.found_user = self.found_user
         self.ids.user_qr_code_screen.found_user = self.found_user
         self.ids.face_recon_screen.found_user = self.found_user
         self.ids.manager.current = "home" # reload the page to trigger his on_pre_enter event
     
     def on_leave(self, *args):
-        self.ids.manager.current = "pin"
+        self.ids.manager.current = "password"
 
 
 
-class UpdateUserPinView(MDScreen):
+class UpdateUserPasswordView(MDScreen):
     found_user = ObjectProperty()
     def __init__(self, **kwargs):
-        super(UpdateUserPinView, self).__init__(**kwargs)
+        super(UpdateUserPasswordView, self).__init__(**kwargs)
         self.__db = Database()
 
     def on_enter(self, *args):
@@ -55,41 +52,72 @@ class UpdateUserPinView(MDScreen):
         if self.found_user:
             print(f"User: {self.found_user}")
             self.__db.db_init(refresh=True)
-            pin = self.__db.get_user_pin(self.found_user)
-            if pin:
-                self.ids.user_pin_field.text = str(pin)
+            password = self.__db.get_user_password(self.found_user)
+            if password:
+                self.ids.user_password_field.text = str(password)
             else:
-                self.ids.user_pin_field.text = "No PIN set yet"
+                self.ids.user_password_field.text = "No Password set yet"
 
-    def update_pin(self):
-        if self.ids.new_pin.ids.password_field.text.strip() and self.ids.confirm_pin.ids.password_field.text.strip():
-            try:
-                new_pin = int(self.ids.new_pin.ids.password_field.text.strip())
-                confirm_pin = int(self.ids.confirm_pin.ids.password_field.text.strip())
-                if new_pin == confirm_pin:
-                    if len(str(new_pin)) == 6 and len(str(confirm_pin)) == 6:
-                        if self.__db.update_user_pin(list(self.found_user), confirm_pin):
-                            toast(f"PIN Successfully updated",
+    def _verify_input_password(self, password:str) -> bool:
+        """
+        Verify if the password meets the rules:
+        at least 6 characters with at least 2 numeric numbers within it
+        :param password:
+        :return:
+        """
+        cnt = 0
+        if len(password) >= 6:
+            for char in password:
+                try:
+                    _ = int(char)
+                    cnt += 1
+                except ValueError:
+                    continue
+            if cnt >=2:
+                return True
+            else:
+                return False
+        else:
+            return False
+
+    def update_password(self):
+        if self.ids.new_password.ids.password_field.text.strip() and self.ids.confirm_password.ids.password_field.text.strip():
+            new_password = self.ids.new_password.ids.password_field.text.strip()
+            confirm_password = self.ids.confirm_password.ids.password_field.text.strip()
+            if new_password == confirm_password:
+                if self._verify_input_password(confirm_password):
+                    if confirm_password not in self.__db.get_db_password_list():
+                        if self.__db.update_user_password(list(self.found_user), confirm_password):
+                            self.ids.user_password_field.text = confirm_password
+                            toast(f"Password Successfully updated",
                                   background=get_color_from_hex(colors["LightGreen"]["500"]), duration=3
                                   )
                             # reset
-                            self.ids.new_pin.ids.password_field.text = ""
-                            self.ids.confirm_pin.ids.password_field.text = ""
-                            self.ids.error_label.text = "* required fields, PIN must have 6 digits"
+                            self.ids.new_password.ids.password_field.text = ""
+                            self.ids.confirm_password.ids.password_field.text = ""
+                            self.ids.error_label.text = "* required fields, Password must have at least 6 characters with a least 2 numeric numbers within it"
                             print("---------")
                             print("Test Update")
                             print("---------")
                             print(self.__db.show_users_table(full=True))
                         else:
-                            toast(f"Error occurred when updating PIN, please try again!",
+                            toast(f"Error occurred when updating Password, please try again!",
                                   background=get_color_from_hex(colors["Red"]["500"]), duration=5
                                   )
+                            # reset
+                            self.ids.new_password.ids.password_field.text = ""
+                            self.ids.confirm_password.ids.password_field.text = ""
                     else:
-                        self.ids.error_label.text = "PIN must have 6 digits"
+                        toast(f"Invalid password, please try another one!",
+                              background=get_color_from_hex(colors["Red"]["500"]), duration=5
+                              )
+                        # reset
+                        self.ids.new_password.ids.password_field.text = ""
+                        self.ids.confirm_password.ids.password_field.text = ""
                 else:
-                    self.ids.error_label.text = "New PIN must be equal to confirmation PIN"
-            except ValueError:
-                self.ids.error_label.text = "New PIN and Confirmation PIN must be numeric numbers"
+                    self.ids.error_label.text = "Password must have at least 6 characters with a least 2 numeric numbers within it"
+            else:
+                self.ids.error_label.text = "New Password must be equal to confirmation Password"
 
         else:
             self.ids.error_label.text = "Please fill required fields"
@@ -115,8 +143,9 @@ class GenerateQRCodeView(MDScreen):
     def generate_qr_code(self):
         # Find & delete old path
         qr_path = self.__db.get_user_qr_img_path(self.found_user)
-        if os.path.exists(qr_path):
-            os.remove(qr_path)
+        if qr_path:
+            if os.path.exists(qr_path):
+                os.remove(qr_path)
 
         qr = qrcode.QRCode(
             version=1,
